Implement PR feedback; fix bugs; move BitmapPool to its own file
